[TST] merged tests for deconcatenate columns
